Refactor seed tags
